refactor(hub): extract generic path utils to kernel/paths (#623)
