+++ packages/hub/src/kernel/collection.ts
@@ -8,7 +8,8 @@ import type { ClassifiedStrategy, ClassifiedVerifyCtx } from '../with-shape/clas
 import type { CrdtMode, CrdtState, LwwMapState, RgaState } from '../with-commit/crdt/crdt.js'
 import type { CrdtStrategy } from '../with-commit/crdt/strategy.js'
 import type { I18nTextDescriptor } from '../shape/via-i18n/core.js'
-import { getAtPath, setAtPathInPlace, stripI18nFilled } from '../shape/via-i18n/core.js'
+import { stripI18nFilled } from '../shape/via-i18n/core.js'
+import { getAtPath, setAtPathInPlace } from './paths.js'
 import type { DictKeyDescriptor, StaticDictDescriptor, DictionaryHandle } from '../shape/via-i18n/dictionary.js'
 import { isStaticDictDescriptor } from '../shape/via-i18n/dictionary.js'
 import { ViaPipeline } from './via-pipeline.js'
