+++ packages/hub/src/kernel/vault.ts
@@ -95,7 +95,7 @@ import { isLinkCollectionName, type LinkSpec, type LinkSetHandle } from '../with
 import { makeLazyLinkSetHandle, type LazyLinkSetHandle } from '../with-shape/links/lazy-handle.js'
 import type { EmbeddingDescriptor } from '../with-lookup/embeddings/index.js'
 import type { I18nTextDescriptor } from '../shape/via-i18n/core.js'
-import { getAtPath } from '../shape/via-i18n/core.js'
+import { getAtPath } from './paths.js'
 import type { ComputedFields } from '../with-formula/computed/index.js'
 import { NO_I18N, type I18nStrategy } from '../shape/via-i18n/strategy.js'
 import { NO_SYNC, type SyncStrategy } from '../with-party/team/sync-strategy.js'
