+++ packages/hub/src/shape/via-i18n/core.ts
@@ -449,61 +449,10 @@ function pickNearestScript(value: Record<string, string>, target: string): strin
 
 // ─── Path helpers (nested i18nFields like 'address.lineOne') ──────────
 
-/**
- * Return all leaf values at `path`, expanding `[].` array wildcards.
- *
- * - `'name'`              → `[obj.name]`
- * - `'address.lineOne'`   → `[obj.address.lineOne]`
- * - `'contacts[].title'`  → `[obj.contacts[0].title, obj.contacts[1].title, …]`
- *
- * Returns an empty array when the path does not resolve (missing key,
- * wrong type, etc.). Used by `enforceI18nOnPut` to validate nested fields.
- */
-export function getAtPath(obj: Record<string, unknown>, path: string): unknown[] {
-  const arrayIdx = path.indexOf('[].')
-  if (arrayIdx !== -1) {
-    const arrayKey = path.slice(0, arrayIdx)
-    const restPath = path.slice(arrayIdx + 3)
-    const arr = obj[arrayKey]
-    if (!Array.isArray(arr)) return []
-    return arr.flatMap(item => {
-      if (!item || typeof item !== 'object' || Array.isArray(item)) return []
-      return getAtPath(item as Record<string, unknown>, restPath)
-    })
-  }
-  const dotIdx = path.indexOf('.')
-  if (dotIdx !== -1) {
-    const head = path.slice(0, dotIdx)
-    const rest = path.slice(dotIdx + 1)
-    const nested = obj[head]
-    if (!nested || typeof nested !== 'object' || Array.isArray(nested)) return []
-    return getAtPath(nested as Record<string, unknown>, rest)
-  }
-  const val = obj[path]
-  return val !== undefined ? [val] : []
-}
-
-/**
- * Mutate `obj` in-place, setting `value` at the nested `path`.
- * Supports dot notation (`'address.lineOne'`) but not array wildcards —
- * auto-translate on `contacts[].title` style paths is not supported.
- */
-export function setAtPathInPlace(
-  obj: Record<string, unknown>,
-  path: string,
-  value: unknown,
-): void {
-  const dotIdx = path.indexOf('.')
-  if (dotIdx !== -1) {
-    const head = path.slice(0, dotIdx)
-    const rest = path.slice(dotIdx + 1)
-    const nested = obj[head]
-    if (!nested || typeof nested !== 'object' || Array.isArray(nested)) return
-    setAtPathInPlace(nested as Record<string, unknown>, rest, value)
-    return
-  }
-  obj[path] = value
-}
+// `getAtPath`/`setAtPathInPlace` are generic dotted-path/[]-wildcard record
+// helpers, not i18n logic — moved to `kernel/paths.ts` (Task 7, #623).
+// Re-exported here for compat with existing importers of this module.
+export { getAtPath, setAtPathInPlace } from '../../kernel/paths.js'
 
 /** Recursively resolve i18nText at a single path within a record copy. */
 function applyAtPath(
